+++ mov/src/org/mov/parser/expression/DayOfWeekExpression.java
@@ -35,7 +35,7 @@ public class DayOfWeekExpression extends TerminalExpression {
         // nothing to do
     }
 
-    public float evaluate(Variables variables, QuoteBundle quoteBundle, Symbol symbol, int day) {
+    public double evaluate(Variables variables, QuoteBundle quoteBundle, Symbol symbol, int day) {
         TradingDate date = quoteBundle.offsetToDate(day);
         return date.getDayOfWeek();
     }
